src/test/java/com/wedeploy/android/OkHttpTransportTest.java
/**
 * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without
 * modification, are permitted provided that the following conditions are met:
 *
 * 1. Redistributions of source code must retain the above copyright notice,
 * this list of conditions and the following disclaimer.
 *
 * 2. Redistributions in binary form must reproduce the above copyright notice,
 * this list of conditions and the following disclaimer in the documentation
 * and/or other materials provided with the distribution.
 *
 * 3. Neither the name of Liferay, Inc. nor the names of its contributors may
 * be used to endorse or promote products derived from this software without
 * specific prior written permission.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
 * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
 * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
 * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
 * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
 * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
 * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
 * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
 * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 * POSSIBILITY OF SUCH DAMAGE.
 */

package com.wedeploy.android;

import com.wedeploy.android.transport.OkHttpTransport;
import com.wedeploy.android.transport.Request;
import com.wedeploy.android.transport.RequestMethod;
import com.wedeploy.android.transport.Response;
import org.junit.Test;

import java.util.concurrent.CountDownLatch;
import java.util.concurrent.TimeUnit;

import static junit.framework.TestCase.assertTrue;
import static org.junit.Assert.fail;

/**
 * @author Silvio Santos
 */
public class OkHttpTransportTest {

	@Test
	public void send() throws InterruptedException {
		OkHttpTransport transport = new OkHttpTransport.Builder().build();

		Request request = new Request.Builder()
			.method(RequestMethod.GET)
			.url("http://www.google.com")
			.build();

		Response response = transport.send(request);

		assertTrue(response.succeeded());
	}

	@Test
	public void sendAsync() throws InterruptedException {
		final CountDownLatch latch = new CountDownLatch(1);

		OkHttpTransport transport = new OkHttpTransport.Builder().build();

		Request request = new Request.Builder()
			.method(RequestMethod.GET)
			.url("http://www.google.com")
			.build();

		transport.sendAsync(request, new Callback() {
			@Override
			public void onSuccess(Response response) {
				assertTrue(response.succeeded());

				latch.countDown();
			}

			@Override
			public void onFailure(Exception e) {
				fail();
			}
		});

		latch.await(5000, TimeUnit.MILLISECONDS);
	}

}
